chore(rurico): migrate to rurico API rev 5569a30
`embed_document` now returns `ChunkedEmbedding`;
`first_chunk()` extracts the first embedding,
enforcing yomu's 1:1 chunk-to-embedding schema.
`run_embed_batch()` consolidates duplicated batch
logic. Renames: `model_paths_if_cached` →
`cached_artifacts`, `download_model` now takes
`ModelId`. `MockEmbedder` updated to
`MockEmbedder::default()`. Types moved to
`storage/types.rs`. `EMBEDDING_DIMS` is now `usize`
natively, removing all `as usize` casts.

Closes #53
