+++ Cargo.toml
@@ -12,7 +12,7 @@ test-support = ["rurico/test-support"]
 [dependencies]
 clap = { version = "4", features = ["derive"] }
 ignore = "0.4"
-rurico = { git = "https://github.com/thkt/rurico", rev = "dd1ad45" }
+rurico = { git = "https://github.com/thkt/rurico", rev = "5569a30" }
 rusqlite = { version = "0.39", features = ["bundled"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -27,7 +27,7 @@ tree-sitter-rust = "0.24"
 tree-sitter-typescript = "0.23"
 
 [dev-dependencies]
-rurico = { git = "https://github.com/thkt/rurico", rev = "dd1ad45", features = ["test-support"] }
+rurico = { git = "https://github.com/thkt/rurico", rev = "5569a30", features = ["test-support"] }
 tempfile = "3"
 
 [profile.release]
